Remove all code related to Direct Deposit v1
Including removing all references to the DD4EDU feature flag
