+++ src/applications/personalization/profile/selectors.js
@@ -1,6 +1,3 @@
-import { toggleValues } from '~/platform/site-wide/feature-toggles/selectors';
-import FEATURE_FLAG_NAMES from '~/platform/utilities/feature-toggles/featureFlagNames';
-
 import {
   cnpDirectDepositBankInfo,
   isEligibleForCNPDirectDeposit,
@@ -79,6 +76,3 @@ export const personalInformationLoadError = state => {
 export const militaryInformationLoadError = state => {
   return state.vaProfile?.militaryInformation?.serviceHistory?.error;
 };
-
-export const showDirectDepositV2 = state =>
-  toggleValues(state)[FEATURE_FLAG_NAMES.directDepositEducation];
