+++ src/applications/personalization/profile/tests/components/Profile.unit.spec.js
@@ -243,6 +243,12 @@ describe('mapStateToProps', () => {
         },
       ],
     },
+    eduPaymentInformation: {
+      accountType: 'Checking',
+      accountNumber: '*****6789',
+      financialInstitutionRoutingNumber: '*****6800',
+      financialInstitutionName: 'Chase Bank',
+    },
   });
   const makeDefaultState = () => ({
     user: {
